requirements.txt
selenium~=4.27.1
pytest~=8.3.4
python-dotenv~=1.0.1
